+++ StreamDeckVSC/Settings/KeySettings.cs
@@ -1,6 +1,0 @@
-﻿namespace StreamDeckVSC.Settings
-{
-    public class KeySettings
-    {
-    }
-}