+++ StreamDeckVSC/Settings/PluginSettings.cs
@@ -1,7 +1,0 @@
-﻿namespace StreamDeckVSC.Settings
-{
-    public class PluginSettings
-    {
-        public bool Feedback { get; set; }
-    }
-}